Create README.md
Welcome to "Blogs by Rohit Gayal," a space for insightful articles, engaging stories, and thought-provoking content. This repository is dedicated to sharing my perspectives and experiences on various topics, such as technology, personal development, travel, and culture.

Whether you're here to learn something new, seek inspiration, or simply enjoy a good read, my blog offers valuable insights and meaningful reflections. Join me on this journey of exploration and discovery as we dive into the world of ideas and creativity.